build(release): Changes needed to get the release running from root.
Modify the build scripts to set things up for releasing the package from root. This means we need to
ignore a bunch of files for npm and clean up the package json in root. Travis will now run
semantic-release from root and run the script to clean package.json before release.