fix(RangeSlider): Remove .css reference
## Summary

Removed reference to `index.css` and removed `index.css` file,
as it does not conform to the rest of the styling patterns.

## Details

Removed the `.css` reference so that webpack is no
longer required to import this repo.

Consolidated the styling rules in the component wrapper
based on what was in the `.css` file.