fix(Radio): Enable keyboard input for radio buttons
- Use opacity instead of display to hide the buttons, which allows access
to the buttons in the dom for keyboard and accessibility use.
- When tabbing to a radio list, the first input is focused. Pressing
down after this causes the second option to be selected. Instead,
this autoselects a value when it is focused, causing the first value
to be selected upon tabbing into a radio list.
- Fix the names of different radio list sets in storybook. The name
attribute must be set correctly when this component is used to get
the correct keyboard behavior.